backend/requirements-dev.txt
# Development dependencies
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
sqlalchemy-stubs==0.4